+++ NEWS.md
@@ -1,6 +1,6 @@
-# mirai 0.11.2.9024 (development)
+# mirai 0.11.2.9025 (development)
 
-* Implements `register()` for registering custom serialization and unserialization functions when using daemons.
+* Implements `serialization()` for registering custom serialization and unserialization functions when using daemons.
 * Introduces `call_mirai_()`, a user-interruptible version of `call_mirai()` suitable for interactive use.
 * Simplification of `mirai()` interface:
   + '.args' will now coerece to a list if an object other than a list is supplied, rather than error.